backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-generativeai==0.3.2
httpx==0.27.2
python-dotenv==1.0.1
sse-starlette==2.1.3
sqlalchemy==2.0.36
typing-extensions==4.12.2
google-cloud-aiplatform==1.42.1
google-auth==2.28.1

# IBM Advanced Components
neo4j==5.14.0  # For Financial Digital Twin
langgraph==0.0.26  # For Autonomous Payment Scheduler  
weaviate-client==3.24.2  # For Knowledge-Augmented LLM
kafka-python==2.0.2  # For Event Streams Risk Guard